+++ src/main/java/com/kyonggi/backend/oauth2/dto/OAuth2Response.java
@@ -1,4 +1,4 @@
-package com.kyonggi.backend.oauth2;
+package com.kyonggi.backend.oauth2.dto;
 
 public interface OAuth2Response {
