+++ src/main/java/com/kyonggi/backend/oauth2/dto/UserDTO.java
@@ -1,4 +1,4 @@
-package com.kyonggi.backend.oauth2;
+package com.kyonggi.backend.oauth2.dto;
 
 import lombok.Getter;
 import lombok.Setter;
